Use uint64_t instead of unsigned
Because SyntaxTreeNode::Type contains over 35 types, we need to use a
64-bit value so that theoretically every type can fit. Not doing this
caused a bug in ParseExpressionOrder where SyntaxTreeNode::Type::value
would be caused as an operator which would cause other problems

+++ cap/FlagList.hh
@@ -1,6 +1,8 @@
 #ifndef FLAG_LIST_HEADER
 #define FLAG_LIST_HEADER
 
+#include <cstdint>
+
 namespace Cap {
 
 template <typename T>
@@ -20,22 +22,22 @@ public:
 
 	constexpr void add(T type)
 	{
-		value |= (1 << static_cast <unsigned> (type));
+		value |= (1UL << static_cast <uint64_t> (type));
 	}
 
 	constexpr bool contains(T type) const
 	{
-		const unsigned result = 1 << static_cast <unsigned> (type);
+		const uint64_t result = 1UL << static_cast <uint64_t> (type);
 		return (value & result) == result;
 	}
 
 	constexpr bool empty() const
 	{
-		return value == 0;
+		return value == 0UL;
 	}
 
 private:
-	unsigned value = 0;
+	uint64_t value = 0UL;
 };
 
 }
